DeepResearcherAgent/requirements.txt
agno
firecrawl
mcp
python-dotenv
streamlit
pydantic
os
asyncio
openai
